+++ sphinx_toolbox/more_autosummary/__init__.py
@@ -322,7 +322,7 @@ class PatchedAutosummary(Autosummary):
 		"""
 		Generate a list of table nodes for the :rst:dir:`autosummary` directive.
 
-		:param items: A list  produced by ``self.get_items``.
+		:param items: A list produced by ``self.get_items``.
 		"""
 
 		table_spec, *other_nodes = super().get_table(items)
@@ -650,6 +650,7 @@ def setup(app: Sphinx) -> SphinxExtMetadata:
 
 	app.setup_extension("sphinx.ext.autosummary")
 	app.setup_extension("autodocsumm")
+	app.setup_extension("sphinx_toolbox.latex")
 
 	app.add_directive("autosummary", PatchedAutosummary, override=True)
 	app.add_directive("autoclasssumm", PatchedAutoDocSummDirective, override=True)